set host and port explicitly, using default Flask ENV variables
This is explicit, so the server and host is less confusing to new students.
The env variables FLASK_RUN_SERVER and FLASK_RUN_PORT are what Flask uses
anyway, instead of using course-specific env variable (which is what we
did at RHUL last year).

The server can replace the defaults in this script when making the
local static zipfile, if needed, because the port and host are going to
be added as config settings on the race server (so, in turn, the task
and tech notes can refer to the editor's host:port with some
confidence.