add minimal (PNG) favicon
This is to remove the potentially confusing repeated requests for favicon.ico,
which appear in the students' editors logs without an obvious reason. Safari
doesn't support PNG favicons but the other browsers do so a lightweight PNG
is the easiest solution.

Favicons are complicated (see: https://dev.to/masakudamatsu/favicon-nightmare-how-to-maintain-sanity-3al7 )
and this is just to mimimise the 404 errors that its absence was causing.