accidentally failed to include the /favicon.png path in base.html
which is why the path was added to app.py, gah!
This _did_ make it into the buggy editor distributed for RHUL 2024–25,
but was erroneously missed out here.

Fixing this now so it can be included in the editor inside the
buggy-race-server repo (as the zip file), not least because we've
just added the tech note about favicon that explictly mentions it.
https://github.com/buggyrace/buggy-race-server/commit/1abe7c0